+++ torch/csrc/Module.cpp
@@ -21,9 +21,10 @@
 #include "torch/csrc/DataLoader.h"
 #include "torch/csrc/Generator.h"
 #include "torch/csrc/autograd/generated/python_nn_functions.h"
-#include "torch/csrc/utils/tensor_dtypes.h"
+#include "torch/csrc/autograd/python_legacy_variable.h"
 #include "torch/csrc/autograd/python_variable.h"
 #include "torch/csrc/tensor/python_tensor.h"
+#include "torch/csrc/utils/tensor_dtypes.h"
 #include "torch/csrc/utils/python_strings.h"
 #include "torch/csrc/utils/tensor_numpy.h"
 #include "torch/csrc/jit/python_tracer.h"
@@ -321,6 +322,14 @@ PyObject *THPModule_setFlushDenormal(PyObject *_unused, PyObject *arg) {
   Py_RETURN_TRUE;
 }
 
+PyObject *THPModule_getDefaultDtype(PyObject *_unused, PyObject *arg) {
+  HANDLE_TH_ERRORS
+  auto dtype = (PyObject*)torch::getDtype(torch::tensor::get_default_tensor_type());
+  Py_INCREF(dtype);
+  return dtype;
+  END_HANDLE_TH_ERRORS
+}
+
 static PyMethodDef TorchMethods[] = {
   {"_initExtension",  (PyCFunction)THPModule_initExtension,   METH_O,       NULL},
   {"_autograd_init",  (PyCFunction)THPAutograd_initExtension, METH_NOARGS,  NULL},
@@ -345,6 +354,7 @@ static PyMethodDef TorchMethods[] = {
   {"_to_dlpack",      (PyCFunction)THPModule_toDLPack,          METH_O,       NULL},
   {"_from_dlpack",    (PyCFunction)THPModule_fromDLPack,        METH_O,       NULL},
   {"set_flush_denormal", (PyCFunction)THPModule_setFlushDenormal, METH_O,     NULL},
+  {"get_default_dtype", (PyCFunction)THPModule_getDefaultDtype, METH_NOARGS,  NULL},
   {NULL, NULL, 0, NULL}
 };
 
@@ -451,6 +461,7 @@ static PyObject* initModule() {
   torch::autograd::initAutogradClosureBindings(module);
   torch::jit::initJITBindings(module);
   torch::autograd::initNNFunctions(module);
+  torch::autograd::init_legacy_variable(module);
 #ifdef WITH_CUDA
   torch::cuda::initModule(module);
 #endif
