+++ torch/csrc/autograd/python_variable.cpp
@@ -17,7 +17,10 @@
 #include "torch/csrc/jit/tracer_state.h"
 #include "torch/csrc/tensor/python_tensor.h"
 #include "torch/csrc/utils/auto_gil.h"
+#include "torch/csrc/utils/cuda_lazy_init.h"
 #include "torch/csrc/utils/python_strings.h"
+#include "torch/csrc/utils/python_arg_parser.h"
+#include "torch/csrc/utils/tensor_new.h"
 
 #include <ATen/ATen.h>
 
@@ -26,6 +29,7 @@
 #include <structmember.h>
 
 using namespace at;
+using namespace torch;
 using namespace torch::autograd;
 
 PyObject *THPVariableClass = nullptr;
@@ -116,80 +120,33 @@ static void THPVariable_dealloc(THPVariable* self)
   Py_TYPE(self)->tp_free((PyObject*)self);
 }
 
-PyObject *THPVariable_pynew(PyTypeObject *type, PyObject *args, PyObject *kwds)
+PyObject *THPVariable_pynew(PyTypeObject *type, PyObject *args, PyObject *kwargs)
 {
   HANDLE_TH_ERRORS
-  THPObjectPtr _data;
-  PyObject *data = nullptr;
-  PyObject *grad_fn = nullptr;
-  char is_volatile = 0;
-  char requires_grad = 0;
-  const char* name = nullptr;
-
-  const char *accepted_args[] = {"data", "requires_grad", "volatile", "_grad_fn", "name", nullptr};
-  if (!PyArg_ParseTupleAndKeywords(args, kwds, "|ObbOz", (char**)accepted_args,
-      &data, &requires_grad, &is_volatile, &grad_fn, &name))
-    return nullptr;
-
-  if (grad_fn == Py_None)
-    grad_fn = nullptr;
-
-  if (is_volatile) {
-    PyErr_WarnEx(PyExc_UserWarning, VOLATILE_WARNING, 1);
+  auto& default_type = torch::tensor::get_default_tensor_type();
+  if (default_type.is_cuda()) {
+    torch::utils::cuda_lazy_init();
   }
-
-  THPUtils_assert(!(is_volatile && requires_grad),
-          "Variable can't be volatile and require_grad at the same time!");
-  THPUtils_assert(!grad_fn || THPFunction_Check(grad_fn),
-          "Variable _grad_fn has to be a Function object or None, but got %s",
-          THPUtils_typename(grad_fn));
-  Tensor tensor;
-  if (!data || data == Py_None) {
-    // For legacy serialization code, create an empty tensor. This is also used
-    // by nn.Parameter() with no arguments.
-    auto var = torch::tensor::get_default_tensor_type().tensor();
-    tensor = static_cast<Variable&>(var).data();
-  } else if (THPVariable_Check(data)) {
-    tensor = ((THPVariable*)data)->cdata.data();
-  } else {
-    throw torch::TypeError("Variable data has to be a tensor, but got %s",
-        THPUtils_typename(data));
-  }
-
-  Variable var;
-  if (grad_fn) {
-    auto grad_fn_ = THPFunction_asFunction((THPFunction*)grad_fn);
-    Edge edge(grad_fn_, grad_fn_->bump_inputs());
-    var = make_variable(std::move(tensor), std::move(edge));
-  } else {
-    var = make_variable(std::move(tensor), requires_grad);
-  }
-
-  if (name) {
-    var.set_name(name);
-  }
-
-  return THPVariable_NewWithVar(type, std::move(var));
+  auto tensor = torch::utils::legacy_tensor_ctor(default_type, args, kwargs);
+  return THPVariable_NewWithVar(type, std::move(tensor));
   END_HANDLE_TH_ERRORS
 }
 
-int THPVariable_pyinit(PyObject *self, PyObject *args, PyObject *kwds)
-{
-  // Ensures that calls to Variable() and subclasses contain data argument.
-  // The 'data' argument is optional in __new__ to handle legacy serialized
-  // Variables.
-  PyObject *data;
-  PyObject *grad_fn = nullptr;
-  char is_volatile = 0;
-  char requires_grad = 0;
-  const char* name = nullptr;
-
-  const char *accepted_args[] = {"data", "requires_grad", "volatile", "_grad_fn", "name", nullptr};
-  if (!PyArg_ParseTupleAndKeywords(args, kwds, "|ObbOz", (char**)accepted_args,
-      &data, &requires_grad, &is_volatile, &grad_fn, &name))
-    return -1;
-
-  return 0;
+static PyObject* THPVariable_make_subclass(PyObject* _ignored, PyObject* args, PyObject* kwargs) {
+  HANDLE_TH_ERRORS
+  static PythonArgParser parser({
+    "_make_subclass(PyObject* cls, Tensor data, bool require_grad=False)",
+  });
+  ParsedArgs<3> parsed_args;
+  auto r = parser.parse(args, kwargs, parsed_args);
+  PyObject* cls = r.pyobject(0);
+  if (!PyType_Check(cls)) {
+    throw TypeError("cls must be a type (got %s)", Py_TYPE(cls)->tp_name);
+  }
+  auto& data = as_variable_ref(r.tensor(1)).data();
+  auto var = make_variable(data, r.toBool(2));
+  return THPVariable_NewWithVar((PyTypeObject*)cls, std::move(var));
+  END_HANDLE_TH_ERRORS
 }
 
 typedef PyObject *(*getter)(PyObject *, void *);
@@ -443,7 +400,7 @@ static struct PyGetSetDef THPVariable_properties[] = {
   {"shape", (getter)THPVariable_get_shape, nullptr, nullptr, nullptr},
   {"is_cuda", (getter)THPVariable_is_cuda, nullptr, nullptr, nullptr},
   {"is_sparse", (getter)THPVariable_is_sparse, nullptr, nullptr, nullptr},
-  {"dtype", (getter)THPVariable_dtype, NULL, NULL, NULL},
+  {"dtype", (getter)THPVariable_dtype, nullptr, nullptr, nullptr},
   {nullptr}
 };
 
@@ -453,9 +410,14 @@ static PyMappingMethods THPVariable_as_mapping = {
   THPVariable_setitem,
 };
 
+static PyMethodDef extra_methods[] = {
+  {"_make_subclass", (PyCFunction)THPVariable_make_subclass, METH_STATIC | METH_VARARGS | METH_KEYWORDS, NULL},
+  {NULL}
+};
+
 PyTypeObject THPVariableType = {
   PyVarObject_HEAD_INIT(nullptr, 0)
-  "torch._C._VariableBase",              /* tp_name */
+  "torch._C._TensorBase",                /* tp_name */
   sizeof(THPVariable),                   /* tp_basicsize */
   0,                                     /* tp_itemsize */
   (destructor)THPVariable_dealloc,       /* tp_dealloc */
@@ -489,7 +451,7 @@ PyTypeObject THPVariableType = {
   0,                                     /* tp_descr_get */
   0,                                     /* tp_descr_set */
   0,                                     /* tp_dictoffset */
-  THPVariable_pyinit,                    /* tp_init */
+  0,                                     /* tp_init */
   0,                                     /* tp_alloc */
   THPVariable_pynew                      /* tp_new */
 };
@@ -505,11 +467,12 @@ bool THPVariable_initModule(PyObject *module)
 {
   static std::vector<PyMethodDef> methods;
   THPUtils_addPyMethodDefs(methods, torch::autograd::variable_methods);
+  THPUtils_addPyMethodDefs(methods, extra_methods);
   THPVariableType.tp_methods = methods.data();
   if (PyType_Ready(&THPVariableType) < 0)
     return false;
   Py_INCREF(&THPVariableType);
-  PyModule_AddObject(module, "_VariableBase", (PyObject *)&THPVariableType);
+  PyModule_AddObject(module, "_TensorBase",   (PyObject *)&THPVariableType);
   torch::autograd::initTorchFunctions(module);
   return true;
 }
